+++ index.js
@@ -3,6 +3,7 @@
 module.exports = {
   rules: {
     'no-let': require('./rules/no-let'),
+    'no-mutation': require('./rules/no-mutation'),
     'no-this': require('./rules/no-this')
   },
   configs: {
@@ -10,6 +11,7 @@ module.exports = {
       rules: {
         'no-var': 'error',
         'fp/no-let': 'error',
+        'fp/no-mutation': 'error',
         'fp/no-this': 'error'
       }
     }